+++ ICD.Common.Utils/EventArguments/BoolEventArgs.cs
@@ -1,4 +1,8 @@
-﻿namespace ICD.Common.Utils.EventArguments
+﻿using System;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Extensions;
+
+namespace ICD.Common.Utils.EventArguments
 {
 	public sealed class BoolEventArgs : GenericEventArgs<bool>
 	{
@@ -18,6 +22,20 @@
 		public BoolEventArgs(BoolEventArgs eventArgs)
 			: this(eventArgs.Data)
 		{
+		}
+	}
+
+	public static class BoolEventArgsExtensions
+	{
+		/// <summary>
+		/// Raises the event safely. Simply skips if the handler is null.
+		/// </summary>
+		/// <param name="extends"></param>
+		/// <param name="sender"></param>
+		/// <param name="data"></param>
+		public static void Raise([CanBeNull]this EventHandler<BoolEventArgs> extends, object sender, bool data)
+		{
+			extends.Raise(sender, new BoolEventArgs(data));
 		}
 	}
 }