+++ ICD.Common.Utils/EventArguments/DateTimeEventArgs.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Extensions;
 
 namespace ICD.Common.Utils.EventArguments
 {
@@ -11,6 +13,20 @@ namespace ICD.Common.Utils.EventArguments
 		public DateTimeEventArgs(DateTime data)
 			: base(data)
 		{
+		}
+	}
+
+	public static class DateTimeEventArgsExtensions
+	{
+		/// <summary>
+		/// Raises the event safely. Simply skips if the handler is null.
+		/// </summary>
+		/// <param name="extends"></param>
+		/// <param name="sender"></param>
+		/// <param name="data"></param>
+		public static void Raise([CanBeNull]this EventHandler<DateTimeEventArgs> extends, object sender, DateTime data)
+		{
+			extends.Raise(sender, new DateTimeEventArgs(data));
 		}
 	}
 }