+++ ICD.Common.Utils/EventArguments/DateTimeNullableEventArgs.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Extensions;
 
 namespace ICD.Common.Utils.EventArguments
 {
@@ -10,6 +12,20 @@ namespace ICD.Common.Utils.EventArguments
 		/// <param name="data"></param>
 		public DateTimeNullableEventArgs(DateTime? data) : base(data)
 		{
+		}
+	}
+
+	public static class DateTimeNullableEventArgsExtensions
+	{
+		/// <summary>
+		/// Raises the event safely. Simply skips if the handler is null.
+		/// </summary>
+		/// <param name="extends"></param>
+		/// <param name="sender"></param>
+		/// <param name="data"></param>
+		public static void Raise([CanBeNull]this EventHandler<DateTimeNullableEventArgs> extends, object sender, DateTime? data)
+		{
+			extends.Raise(sender, new DateTimeNullableEventArgs(data));
 		}
 	}
 }