+++ ICD.Common.Utils/EventArguments/GenericEventArgs.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Extensions;
 
 namespace ICD.Common.Utils.EventArguments
 {
@@ -23,6 +25,20 @@ namespace ICD.Common.Utils.EventArguments
 		public GenericEventArgs(T data)
 		{
 			m_Data = data;
+		}
+	}
+
+	public static class GenericEventArgsExtensions
+	{
+		/// <summary>
+		/// Raises the event safely. Simply skips if the handler is null.
+		/// </summary>
+		/// <param name="extends"></param>
+		/// <param name="sender"></param>
+		/// <param name="data"></param>
+		public static void Raise<T>([CanBeNull]this EventHandler<GenericEventArgs<T>> extends, object sender, T data)
+		{
+			extends.Raise(sender, new GenericEventArgs<T>(data));
 		}
 	}
 }