+++ ICD.Common.Utils/EventArguments/UShortEventArgs.cs
@@ -1,4 +1,6 @@
-﻿using ICD.Common.Properties;
+﻿using System;
+using ICD.Common.Properties;
+using ICD.Common.Utils.Extensions;
 
 namespace ICD.Common.Utils.EventArguments
 {
@@ -11,6 +13,20 @@ namespace ICD.Common.Utils.EventArguments
 		/// <param name="data"></param>
 		public UShortEventArgs(ushort data) : base(data)
 		{
+		}
+	}
+
+	public static class UShortEventArgsExtensions
+	{
+		/// <summary>
+		/// Raises the event safely. Simply skips if the handler is null.
+		/// </summary>
+		/// <param name="extends"></param>
+		/// <param name="sender"></param>
+		/// <param name="data"></param>
+		public static void Raise([CanBeNull]this EventHandler<UShortEventArgs> extends, object sender, ushort data)
+		{
+			extends.Raise(sender, new UShortEventArgs(data));
 		}
 	}
 }